Add crf_util.py
A place for functions useful to many workflows.
Update gridsearch_train_crfs to use crf_util.
